Make COSMOVISOR_ENABLED and START_CMD work as intended
Teach the run.sh to use START_CMD if set, even when SNAPSHOT_PATH is unused
(previously, it would exec whatever command line passed to the docker, or do
nothing if none).

This also fixes the case of COSMOVISOR_ENABLED=1, which sets START_CMD to
start cosmovisor but then was not actually run before.